changed type of lower and upper bound of unsigned int parameters to size_t
it used to be int, which makes little sense, the parameter value is not signed
and so its range is not signed either.

+++ include/General/Parameter.h
@@ -266,9 +266,9 @@ class Parameter
     /// Even if the range is an interval, here all numbers are listed
     std::set<size_t> unsigned_range;
     /// @brief the range is the interval [unsigned_min, unsigned_max]
-    int unsigned_min;
+    size_t unsigned_min;
     /// @brief the range is the interval [unsigned_min, unsigned_max]
-    int unsigned_max;
+    size_t unsigned_max;
     /// @brief the range is the interval [min, max]
     double double_min;
     /// @brief the range is the interval [min, max]
